requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-core
langchain-classic
chromadb
langchain-google-genai
google-generativeai
python-dotenv
aiofiles
